feat: replace pricing and selection heuristics with exact cost minimization
Pricing (cost horizon):
- Price terminal late work and terminal replica time in closed form.
- Close scenarios only after all modeled events, including every
  supplied future release.
- Aggregate overdue work without loss and remove the backlog age split.
- Replace epsilon ledger gates and completion cutoffs with exact
  exhaustion branches.
- Remove the fixed schedule-visibility clip, the zero-continuation
  branch, and the breakpoint pass limit.

Selection:
- Select the exact cost argmin over paired posterior samples.
- Delete the bootstrap classifier, the pilot pass, the incumbency
  preference, the mean transition price, the posterior knee cap, the
  selection clamp, deterministic placement scoring, the posterior-mean
  partition activation, and the active-partition action cap.
- Charge replica time by actual membership at each instant.
- Price empty-member departures movement-free at request time.
- Repair failed members symmetrically; the optimistic bias of the
  sampled-truth successor is documented at the site.
- Report the top-two paired standard error as a diagnostic only.

Core suite: 98/98 with the simulation profile. Known open item: the
debug profile trips a new debug assertion in edf.rs; root cause runs
next. Regime suites intentionally not run: they wait until every audit
finding is fixed.

+++ crates/scale-core/benches/step.rs
@@ -367,14 +367,14 @@ fn populate_staggered(
     observation: &mut ObservationBuffer,
     case: BenchmarkCase,
 ) -> Result<(), BenchmarkError> {
-    const HORIZON_SECONDS: u64 = 30;
-    let events_per_cohort = f64::from(case.offered_events_per_second) * HORIZON_SECONDS as f64
+    const HORIZON_SECONDS: u32 = 30;
+    let events_per_cohort = f64::from(case.offered_events_per_second) * f64::from(HORIZON_SECONDS)
         / f64::from(case.cohort_count_max);
     for cohort in 0..case.cohort_count_max {
         let release_micros = u64::from(cohort) * 100_000;
         observation.push_cohort(Cohort {
             release_micros,
-            deadline_micros: release_micros + HORIZON_SECONDS * 1_000_000,
+            deadline_micros: release_micros + u64::from(HORIZON_SECONDS) * 1_000_000,
             offered_events: events_per_cohort,
             partition: cohort % 64,
             demand_class: DemandClass::Normal,
